+++ samplesheets/tests/test_permissions_api.py
@@ -1,10 +1,13 @@
 """Tests for REST API View permissions in the samplesheets app"""
 
 from django.urls import reverse
+
 from projectroles.models import SODAR_CONSTANTS
 from projectroles.tests.test_models import RemoteSiteMixin, RemoteProjectMixin
 from projectroles.tests.test_permissions import TestProjectPermissionBase
+from projectroles.tests.test_permissions_api import TestProjectAPIPermissionBase
 
+from samplesheets.models import Investigation
 from samplesheets.tests.test_io import SampleSheetIOMixin
 from samplesheets.tests.test_permissions import (
     SHEET_PATH,
@@ -19,7 +22,7 @@ class TestInvestigationRetrieveAPIView(
     SampleSheetIOMixin,
     RemoteSiteMixin,
     RemoteProjectMixin,
-    TestProjectPermissionBase,
+    TestProjectAPIPermissionBase,
 ):
     """Tests for InvestigationRetrieveAPIView permissions"""
 
@@ -44,9 +47,77 @@ class TestInvestigationRetrieveAPIView(
             self.contributor_as.user,
             self.guest_as.user,
         ]
-        self.assert_response(url, good_users, status_code=200)
-        self.assert_response(url, [self.user_no_roles], status_code=403)
-        self.assert_response(url, [self.anonymous], status_code=401)
+        self.assert_response_api(url, good_users, 200)
+        self.assert_response_api(url, [self.user_no_roles], 403)
+        self.assert_response_api(url, [self.anonymous], 401)
+
+
+class TestSampleSheetImportAPIView(
+    SampleSheetIOMixin,
+    RemoteSiteMixin,
+    RemoteProjectMixin,
+    TestProjectAPIPermissionBase,
+):
+    """Tests for SampleSheetImportAPIView permissions"""
+
+    def setUp(self):
+        super().setUp()
+
+        self.zip_file = open(SHEET_PATH, 'rb')
+        self.post_data = {'file': self.zip_file}
+
+    def tearDown(self):
+        self.zip_file.close()
+        super().tearDown()
+
+    def test_post(self):
+        """Test post() in SampleSheetImportAPIView"""
+        url = reverse(
+            'samplesheets:api_import',
+            kwargs={'project': self.project.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+            self.owner_as.user,
+            self.delegate_as.user,
+            self.contributor_as.user,
+        ]
+        bad_users = [
+            self.guest_as.user,
+            self.user_no_roles,
+        ]
+
+        def _cleanup():
+            self.zip_file.seek(0)
+            Investigation.objects.filter(project=self.project).delete()
+
+        self.assert_response_api(
+            url,
+            good_users,
+            status_code=200,
+            method='POST',
+            format='multipart',
+            data=self.post_data,
+            cleanup_method=_cleanup,
+        )
+        self.assert_response_api(
+            url,
+            bad_users,
+            status_code=403,
+            method='POST',
+            format='multipart',
+            data=self.post_data,
+            cleanup_method=_cleanup,
+        )
+        self.assert_response_api(
+            url,
+            [self.anonymous],
+            status_code=401,
+            method='POST',
+            format='multipart',
+            data=self.post_data,
+            cleanup_method=_cleanup,
+        )
 
 
 class TestRemoteSheetGetAPIView(